Move POD_TEMPLATES_TYPES to avoid circular dependencies.
